refactor: move fetchSticker RPC into signal-enhancements module
